Update FB SMA-Price Crossover Research
Compared the cross between price and 25 day, 50 day, and 75 day SMA. Found that among those, the 50 day SMA seems to be the best when analyzing price/SMA crossovers. When the price crosses above the SMA, it is buy opportunity, and vice versa. The 50 day SMA is not subject to as much flip-flopping as a 25 day SMA, and also is a bit more quick to react to changes than the 75 day SMA which makes it a suitable SMA for buying/selling indicators.